+++ src/spells/blocks.cpp
@@ -1121,6 +1121,9 @@ public:
 		return {};
 	}
 
+	static void fixFO76ShaderPropertyName( NifModel * nif, char * blockData, const QModelIndex & iBlock,
+											const QString & blockType, const QStringList & strings );
+
 	bool isApplicable( const NifModel * nif, const QModelIndex & index ) override final
 	{
 		Q_UNUSED( index );
@@ -1165,6 +1168,8 @@ public:
 							ds >> strings;
 
 						QModelIndex block = nif->insertNiBlock( bType, nif->getBlockCount() );
+						if ( buffer.pos() <= ( data.size() - 4 ) )
+							fixFO76ShaderPropertyName( nif, data.data() + buffer.pos(), block, bType, strings );
 						nif->loadIndex( buffer, block );
 						blockLink( nif, index, block );
 
@@ -1189,6 +1194,23 @@ public:
 		return QModelIndex();
 	}
 };
+
+void spPasteBlock::fixFO76ShaderPropertyName( NifModel * nif, char * blockData, const QModelIndex & iBlock,
+												const QString & blockType, const QStringList & strings )
+{
+	if ( nif->getBSVersion() < 151 )
+		return;
+	if ( !( blockType == "BSLightingShaderProperty" || blockType == "BSEffectShaderProperty" ) )
+		return;
+	// hack to work around issues with pasting Fallout 76 and Starfield shader property blocks
+	// where the data is conditional based on the block name being empty
+	QModelIndex	iName = nif->getIndex( iBlock, "Name" );
+	if ( iName.isValid() && !strings.isEmpty() ) {
+		nif->set<QString>( iName, strings.first() );
+		// write the new string index to the buffer (FIXME: this may be non-portable)
+		*( reinterpret_cast< std::int32_t * >( blockData ) ) = nif->get<qint32>( iName );
+	}
+}
 
 REGISTER_SPELL( spPasteBlock )
 
@@ -1258,6 +1280,8 @@ public:
 						if ( nif->checkVersion( 0x14010001, 0 ) )
 							ds >> strings;
 
+						if ( buffer.pos() <= ( data.size() - 4 ) )
+							spPasteBlock::fixFO76ShaderPropertyName( nif, data.data() + buffer.pos(), index, bType, strings );
 						nif->loadIndex( buffer, index );
 
 						// NiDataStream RTTI arg values
@@ -1472,6 +1496,8 @@ QModelIndex spPasteBranch::cast( NifModel * nif, const QModelIndex & index )
 						bType = nif->extractRTTIArgs( bType, metadata );
 
 						QModelIndex block = nif->insertNiBlock( bType, -1 );
+						if ( buffer.pos() <= ( data.size() - 4 ) )
+							spPasteBlock::fixFO76ShaderPropertyName( nif, data.data() + buffer.pos(), block, bType, strings );
 						if ( !nif->loadAndMapLinks( buffer, block, blockMap ) )
 							return index;